reorganize exports. simplify 'as-a-library' call types
