+++ bindings/cloudflare/src/cloudflare_worker.ts
@@ -20,12 +20,11 @@
 
 type Env = any;
 
-import { FetchApiRequest, FetchApiResponse } from "@google-cloud/recaptcha-edge";
 import { CloudflareContext, processRequest, recaptchaConfigFromEnv } from "./index";
 
 export default {
   async fetch(request: Request, env: Env, ctx: ExecutionContext): Promise<Response> {
     const cfctx = new CloudflareContext(env, ctx, recaptchaConfigFromEnv(env));
-    return processRequest(cfctx, new FetchApiRequest(request)).then((v) => (v as FetchApiResponse).asResponse());
+    return processRequest(cfctx, request);
   },
 } satisfies ExportedHandler;
