+++ bindings/cloudflare/src/wrappers.ts
@@ -1,0 +1,47 @@
+/**
+ * Copyright 2024 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+// eslint-disable-next-line  @typescript-eslint/no-unused-vars
+import {
+  callCreateAssessment,
+  callListFirewallPolicies,
+  processRequest as baseProcessRequest,
+  Event,
+  FetchApiRequest,
+  FetchApiResponse,
+  Assessment,
+  ListFirewallPoliciesResponse,
+} from "@google-cloud/recaptcha-edge";
+
+import { CloudflareContext } from "./context";
+
+export async function processRequest(ctx: CloudflareContext, r: Request): Promise<Response> {
+  const v = await baseProcessRequest(ctx, new FetchApiRequest(r));
+  return (v as FetchApiResponse).asResponse();
+}
+
+export function createAssessment(
+  ctx: CloudflareContext,
+  r: Request,
+  environment?: [string, string],
+  additionalParams?: Event,
+): Promise<Assessment> {
+  return callCreateAssessment(ctx, new FetchApiRequest(r), environment, additionalParams);
+}
+
+export function listFirewallPolicies(ctx: CloudflareContext): Promise<ListFirewallPoliciesResponse> {
+  return callListFirewallPolicies(ctx);
+}
